backend/requirements.txt
flask==3.0.3
flask-wtf==1.2.1
flask-cors==4.0.1
flask-sqlalchemy==3.1.1